dependencies for viewModel, added internet permission, updated endpoint url

+++ app/src/main/java/ua/honchar/reddit/domain/model/PostModelView.kt
@@ -8,6 +8,6 @@ data class PostModelView(
     val thumbnail: String,
     val numComments: Int,
     val created: Long,
-    val ratio: Int,
+    val ratio: Double,
     val ups: Int
 )
